Add simple support for creating and storing 'apps'.
Apps are server side recordings of client apps that are
going to use resources on this server, for the sake of
the users of those client apps.

An app is represented as a tiddler in a bag (defaults to
oauth_apps). The title is the client id. Additional fields
are kept on the tiddler (see the tests).

Also: move some of the oauth2client handling to its own module.
That stuff will be resurrected once the server is more complete.

The next step is handling auth url and token urls.