Provide an 'authorize' endpoint.
This is the handler that is the human interaction with a provider.
The human authorizes, at the provider, that the consumer can
consume the scopes provided.

This does that authorization step, returning a 'code' to the
callback_url on the consumer.

For these tests the provider and the consumer are the same
place.

This is not fully robust just yet. No scope validation is being
done, and methods that check to see if this user has already
authorized this consumer in the recent past are not implemented.

I'm driving this development via client implementation. That is
i'm fulfilling the server side requirements that a client demands.
There will likely be blanks that need to be filled in later.

However, it is interesting to see it developing, because in a
way this is feeling a lot like privateer: The consumer is being given
a token that on the provider is decoded to indicate what it is allowed
to do, for who.

One major issue that will need to be dealt with: None of this stuff
is particularly secure if not done over TLS. Which means certs.