Add support for making an access token and then using it in an extractor.
We now have a complete system where we can go through the full
process of authorizing a client, getting a code, use that to
get an access token, and then accessing a restricted resource
with that token.

That's pretty cool but there are several things missing:

* handling of 'scopes'
* managing the 'state' field that should be passed around
  in requests
* legible code
* more tests from more angles

But this is pretty good progress. Need to do some analysis
and auditing of the code to make sure it is actually right.

Also: Bearer tokens (what this uses in the extractor) are
no good unless over SSL on which the certs are validated!

(Just throwing that in there as a reminder)