Improve auth handling when getting a token.
After reading the whole RFC it's become clear there are
some naming conflicts between the in the field names that are
being used (some held over from oauth1) and those names used
in the rfc. So, for example, consumer and provider are no
longer quite right, because the "provider" is now split into
an auth server and a resource server. I should probably endeavor
to clean up the module names (again).